pyproject.toml
[project]
name = "starlink-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydoover>=0.4.11",
    "transitions>=0.9.2",
    "starlink-grpc-core",
]

[project.scripts]
doover-app-run = "starlink_manager:main"
export-config = "starlink_manager.app_config:export"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "aiohttp>=3.11.18",
    "grpcio>=1.71.0",
    "protobuf>=6.30.2",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "requests>=2.32.3",
]
